refactor: move some logic
