+++ crates/operator/Cargo.toml
@@ -18,6 +18,7 @@ futures-util.workspace = true
 # eigen-rs
 eigen-aggregator.workspace = true
 eigen-client-avsregistry.workspace = true
+eigen-client-elcontracts.workspace = true
 eigen-crypto-bls.workspace = true
 eigen-task-manager.workspace = true
 eigen-types.workspace = true
