+++ crates/operator/src/error.rs
@@ -1,4 +1,6 @@
+use alloy::signers::local::LocalSignerError;
 use eigen_client_avsregistry::error::AvsRegistryError;
+use eigen_client_elcontracts::error::ElContractsError;
 // use eigen_config::error::ConfigError;
 use eigen_crypto_bls::error::BlsError;
 use eigen_task_manager::{event_decoder::AbiDecodeError, TaskManagerError};
@@ -52,4 +54,10 @@ pub enum OperatorError {
     /// Invalid deposit tokens
     #[error("Invalid deposit tokens")]
     InvalidDepositTokens,
+    /// ELChainWriter Error
+    #[error("ElContractsError Error")]
+    ElContractsError(#[from] ElContractsError),
+    /// Failed to decrypt keystore
+    #[error("Failed to decrypt keystore")]
+    SignerError(#[from] LocalSignerError),
 }
