+++ tests/testing-utils/Cargo.toml
@@ -18,8 +18,6 @@ eigen-client-elcontracts.workspace = true
 eigen-common.workspace = true
 eigen-utils.workspace = true
 eigen-crypto-bls.workspace = true
-eigen-operator.workspace = true
-eigen-types.workspace = true
 eigen-task-manager.workspace = true
 rand.workspace = true
 serde.workspace = true
